requirements.txt
#backports.zoneinfo==0.2.1
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
importlib-metadata==4.10.0
Markdown==3.3.6
mysqlclient==2.1.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
zipp==3.7.0
gunicorn==20.1.0
drf-nested-routers==0.93.4
ipython==8.0.1
